add config, services, separate routes
